@@ -1,47 +1,44 @@
 {
-    "name": "carriage-web",
-    "version": "1.0.0",
-    "description": "",
-    "main": "index.js",
-    "dependencies": {
-        "@types/jquery": "^3.3.33",
-        "babel-eslint": "^10.1.0",
-        "eslint": "^6.8.0",
-        "eslint-config-airbnb-base": "^14.0.0",
-        "eslint-config-react-app": "^5.2.0",
-        "eslint-plugin-flowtype": "^3.13.0",
-        "eslint-plugin-import": "^2.20.1",
-        "eslint-plugin-jsx-a11y": "^6.2.3",
-        "eslint-plugin-promise": "^4.2.1",
-        "eslint-plugin-react": "^7.18.3",
-        "express": "^4.17.1",
-        "google-auth-library": "^6.0.0",
-        "react-google-login": "^5.1.3",
-        "typescript": "^4.0.3"
-    },
-    "devDependencies": {
-        "@typescript-eslint/eslint-plugin": "^2.18.0",
-        "@typescript-eslint/parser": "^2.18.0",
-        "eslint-plugin-react-hooks": "^1.7.0"
-    },
-    "scripts": {
-        "heroku-postbuild": "npm run build:frontend && npm run build:server",
-        "build:frontend": "cd frontend && npm run build && cd ..",
-        "build:server": "cd server && npm run tsc && cd ..",
-        "install:frontend": "cd frontend && npm install && cd ..",
-        "install:server": "cd server && npm install && cd ..",
-        "postinstall": "npm run install:frontend && npm run install:server",
-        "lint": "eslint . --ext .js --ext .jsx --ext .ts --ext .tsx --fix",
-        "test": "echo \"Error: no test specified\" && exit 1"
-    },
-    "repository": {
-        "type": "git",
-        "url": "git+https://github.com/cornell-dti/carriage-web.git"
-    },
-    "author": "",
-    "license": "ISC",
-    "bugs": {
-        "url": "https://github.com/cornell-dti/carriage-web/issues"
-    },
-    "homepage": "https://github.com/cornell-dti/carriage-web#readme"
+  "name": "carriage-web",
+  "version": "1.0.0",
+  "description": "",
+  "main": "index.js",
+  "dependencies": {},
+  "devDependencies": {
+    "@babel/core": "^7.12.16",
+    "@typescript-eslint/eslint-plugin": "^2.18.0",
+    "@typescript-eslint/parser": "^2.18.0",
+    "babel-eslint": "^10.1.0",
+    "eslint": "^6.8.0",
+    "eslint-config-airbnb-base": "^14.2.1",
+    "eslint-plugin-import": "^2.22.1",
+    "eslint-plugin-promise": "^4.3.1",
+    "eslint-plugin-react": "^7.22.0",
+    "eslint-plugin-react-hooks": "^1.7.0",
+    "typescript": "^4.1.5"
+  },
+  "scripts": {
+    "heroku-postbuild": "npm run build:frontend && npm run build:server",
+    "start:frontend": "cd frontend && npm run start",
+    "start:server": "cd server && npm run start",
+    "build:frontend": "cd frontend && npm run build",
+    "build:server": "cd server && npm run tsc",
+    "install:frontend": "cd frontend && npm install",
+    "install:server": "cd server && npm install",
+    "postinstall": "npm run install:frontend && npm run install:server",
+    "quiet-lint": "eslint . --ext .js --ext .jsx --ext .ts --ext .tsx --fix --quiet",
+    "lint": "eslint . --ext .js --ext .jsx --ext .ts --ext .tsx --fix",
+    "type-check": "cd frontend && npm run type-check && cd ../server && npm run type-check",
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/cornell-dti/carriage-web.git"
+  },
+  "author": "",
+  "license": "ISC",
+  "bugs": {
+    "url": "https://github.com/cornell-dti/carriage-web/issues"
+  },
+  "homepage": "https://github.com/cornell-dti/carriage-web#readme"
 }
