@@ -1,4 +1,4 @@
-import React, { useState, FunctionComponent } from 'react';
+import React, { useState, ReactNode } from 'react';
 import { GoogleLogin } from 'react-google-login';
 import { useHistory, useLocation } from 'react-router-dom';
 import ReqContext from '../../context/req';
@@ -6,8 +6,11 @@ import useClientId from '../../hooks/useClientId';
 import AuthContext from '../../context/auth';
 import LandingPage from '../../pages/Landing/Landing';
 
+type AuthManagerProps = {
+  children: ReactNode;
+};
 
-export const AuthManager: FunctionComponent = ({ children }) => {
+export const AuthManager = ({ children }: AuthManagerProps) => {
   const [signedIn, setSignedIn] = useState(false);
   const [jwt, setJWT] = useState('');
   const clientId = useClientId();
@@ -39,16 +42,18 @@ export const AuthManager: FunctionComponent = ({ children }) => {
     const { id_token: token } = googleUser.getAuthResponse();
     const { email } = googleUser.profileObj;
 
-    const serverJWT = await fetch('/api/auth', withDefaults({
-      method: 'POST',
-      body:
-        JSON.stringify({
+    const serverJWT = await fetch(
+      '/api/auth',
+      withDefaults({
+        method: 'POST',
+        body: JSON.stringify({
           token,
           email,
           table: 'Dispatchers',
           clientId,
         }),
-    }))
+      }),
+    )
       .then((res) => res.json())
       .then((json) => json.jwt);
 
@@ -77,7 +82,7 @@ export const AuthManager: FunctionComponent = ({ children }) => {
         onSuccess={onSignIn}
         onFailure={(err) => console.error(err)}
         clientId={clientId}
-        cookiePolicy='single_host_origin'
+        cookiePolicy="single_host_origin"
         isSignedIn
       />
       <LandingPage />
