@@ -9,8 +9,12 @@ type WeekState = {
 };
 
 const initialState: WeekState = {
-  selectDay: () => { },
-  deselectDay: () => { },
+  selectDay: () => {
+    // do nothing
+  },
+  deselectDay: () => {
+    // do nothing
+  },
   isDayOpen: () => false,
   isDaySelectedByInstance: () => false,
   getSelectedDays: () => [],
@@ -22,7 +26,7 @@ export const useWeek = () => React.useContext(WeekContext);
 
 type WeekProviderProps = {
   children: React.ReactNode;
-}
+};
 
 type WeekType = {
   [day: string]: number;
@@ -52,13 +56,9 @@ export const WeekProvider = ({ children }: WeekProviderProps) => {
 
   const isDayOpen = (day: string) => week[day] === -1;
 
-  const isDaySelectedByInstance = (day: string, index: number) => (
-    week[day] === index
-  );
+  const isDaySelectedByInstance = (day: string, index: number) => week[day] === index;
 
-  const getSelectedDays = (index: number) => (
-    Object.keys(week).filter((day) => week[day] === index)
-  );
+  const getSelectedDays = (index: number) => Object.keys(week).filter((day) => week[day] === index);
 
   return (
     <WeekContext.Provider
