@@ -42,7 +42,7 @@ const RiderModalInfo = ({ onSubmit }: ModalFormProps) => {
             placeholder="Name"
             ref={register({ required: true, pattern: /^[a-zA-Z]+\s[a-zA-Z]+/ })}
           />
-          {errors.email && "enter a valid name"}
+          {errors.email && 'enter a valid name'}
         </div>
         <div className={cn(styles.gridR1, styles.gridC2)}>
           <Input
@@ -51,7 +51,7 @@ const RiderModalInfo = ({ onSubmit }: ModalFormProps) => {
             placeholder="NetID"
             ref={register({ required: true, pattern: /^[a-zA-Z]+[0-9]+$/ })}
           />
-          {errors.email && "enter a valid netid"}
+          {errors.email && 'enter a valid netid'}
         </div>
         <div className={cn(styles.gridR2, styles.gridC1)}>
           <Input
@@ -60,7 +60,7 @@ const RiderModalInfo = ({ onSubmit }: ModalFormProps) => {
             placeholder="Email"
             ref={register({ required: true, pattern: /^\w+([.-]?\w+)*@\w+([.-]?\w+)*(\.\w{2,3})+$/ })}
           />
-          {errors.email && "enter a valid email"}
+          {errors.email && 'enter a valid email'}
         </div>
         <div className={cn(styles.gridR2, styles.gridC2)}>
           <Input
@@ -69,7 +69,7 @@ const RiderModalInfo = ({ onSubmit }: ModalFormProps) => {
             placeholder="Phone Number"
             ref={register({ required: true, pattern: /^[0-9]{10}$/ })}
           />
-          {errors.phoneNumber && "enter a valid phone number"}
+          {errors.phoneNumber && 'enter a valid phone number'}
         </div>
         <div className={cn(styles.gridR3, styles.gridC1)}>
           <Input
@@ -78,7 +78,7 @@ const RiderModalInfo = ({ onSubmit }: ModalFormProps) => {
             placeholder="Needs"
             ref={register({ required: true })}
           />
-          {errors.email && "enter some needs"}
+          {errors.email && 'enter some needs'}
         </div>
         <div className={cn(styles.gridR3, styles.gridC2)}>
           <Input
@@ -87,7 +87,7 @@ const RiderModalInfo = ({ onSubmit }: ModalFormProps) => {
             placeholder="Address"
             ref={register({ required: true })}
           />
-          {errors.email && "enter an address"}
+          {errors.email && 'enter an address'}
         </div>
         <div className={cn(styles.gridR4, styles.gridC1)}>
           <Label htmlFor="start">Start Date:</Label>
