@@ -5,19 +5,23 @@ import DropDownInput from './DropDownInput';
 import CheckBoxInput from './CheckBoxInput';
 
 type FormProps = {
-  onClick: ((newRider: Rider) => void);
-}
+  onClick: (newRider: Rider) => void;
+};
+
+const emptyStringArray: string[] = [];
 
 const Form = (props: FormProps) => {
   const today = new Date();
-  const date = `${today.getMonth() + 1}/${today.getDate()}/${today.getFullYear()}`;
+  const date = `${
+    today.getMonth() + 1
+  }/${today.getDate()}/${today.getFullYear()}`;
   const [newRider, setNewRider] = useState({
     id: '',
     firstName: '',
     lastName: '',
     phoneNumber: '',
     email: '',
-    accessibilityNeeds: new Array<string>(),
+    accessibilityNeeds: emptyStringArray,
     description: '',
     joinDate: date,
     pronouns: 'she/her/hers',
@@ -51,7 +55,7 @@ const Form = (props: FormProps) => {
       }
     } else if (fieldName === 'phone') {
       const phoneFormat = /^[0-9]{10}$/;
-      if ((fieldValue.length > 0) && fieldValue.match(phoneFormat)) {
+      if (fieldValue.length > 0 && fieldValue.match(phoneFormat)) {
         newRider.phoneNumber = fieldValue;
         setValidPhone(true);
       } else {
@@ -60,7 +64,7 @@ const Form = (props: FormProps) => {
       }
     } else if (fieldName === 'email') {
       const netIdFormat = /^[a-zA-Z]+[0-9]+$/;
-      if ((fieldValue.length > 0) && fieldValue.match(netIdFormat)) {
+      if (fieldValue.length > 0 && fieldValue.match(netIdFormat)) {
         newRider.email = fieldValue.concat('@cornell.edu');
         setValidEmail(true);
       } else {
@@ -88,7 +92,8 @@ const Form = (props: FormProps) => {
     } else {
       if (fieldName === 'needWheel') {
         newRider.accessibilityNeeds.push('Wheelchair');
-      } if (fieldName === 'needCrutches') {
+      }
+      if (fieldName === 'needCrutches') {
         newRider.accessibilityNeeds.push('Crutches');
       } else if (fieldName === 'needAssist') {
         newRider.accessibilityNeeds.push('Assistant');
@@ -98,9 +103,15 @@ const Form = (props: FormProps) => {
   };
   const handleSubmit = (evt: any) => {
     evt.preventDefault();
-    const validRider = validFirstName && validLastName && validPhone
-      && validEmail && validDesc && validAddress;
-    if (validRider) { props.onClick(newRider); }
+    const validRider = validFirstName
+      && validLastName
+      && validPhone
+      && validEmail
+      && validDesc
+      && validAddress;
+    if (validRider) {
+      props.onClick(newRider);
+    }
   };
   return (
     <>
@@ -112,30 +123,38 @@ const Form = (props: FormProps) => {
             labelText="First Name: "
             feedback="Please enter a first name"
             showFormFeedback={validFirstName}
-            handleInput={(e) => handleInput(e)} />
+            handleInput={(e) => handleInput(e)}
+          />
           <TextInput
             labelName="lastName"
             labelText="Last Name: "
             feedback="Please enter a last name"
             showFormFeedback={validLastName}
-            handleInput={(e) => handleInput(e)} />
+            handleInput={(e) => handleInput(e)}
+          />
           <TextInput
             labelName="phone"
             labelText="Phone Number: "
             feedback="Enter a phone number in the form xxx-xxx-xxxx"
             showFormFeedback={validPhone}
-            handleInput={(e) => handleInput(e)} />
+            handleInput={(e) => handleInput(e)}
+          />
           <TextInput
             labelName="email"
             labelText="NetId: "
             feedback="Enter a valid netid"
             showFormFeedback={validEmail}
-            handleInput={(e) => handleInput(e)} />
+            handleInput={(e) => handleInput(e)}
+          />
           <CheckBoxInput
             labelText="Accessibility Needs: "
             checkboxId="accessibility"
-            options={["needWheel", "needCrutches", "needAssist"]}
-            optionLabels={["Needs Wheelchair", "Has Crutches", "Needs Assistant"]}
+            options={['needWheel', 'needCrutches', 'needAssist']}
+            optionLabels={[
+              'Needs Wheelchair',
+              'Has Crutches',
+              'Needs Assistant',
+            ]}
             handleInput={handleInput}
           />
           <TextInput
@@ -143,18 +162,21 @@ const Form = (props: FormProps) => {
             labelText="Description: "
             feedback="Enter a description"
             showFormFeedback={validDesc}
-            handleInput={(e) => handleInput(e)} />
+            handleInput={(e) => handleInput(e)}
+          />
           <DropDownInput
             labelName="pronouns"
             labelText="Pronouns: "
-            options={["she/her/hers", "he/him/his", "neutral"]}
-            handleInput={(e) => handleInput(e)} />
+            options={['she/her/hers', 'he/him/his', 'neutral']}
+            handleInput={(e) => handleInput(e)}
+          />
           <TextInput
             labelName="address"
             labelText="Address: "
             feedback="Enter an address"
             showFormFeedback={validAddress}
-            handleInput={(e) => handleInput(e)} />
+            handleInput={(e) => handleInput(e)}
+          />
           <input type="submit" value="Submit" />
         </form>
       </div>
