@@ -3,11 +3,11 @@ import Form from '../UserForms/DriversForm';
 import styles from './table.module.css';
 
 type Driver = {
-  name: string,
-  netid: string,
-  email: string,
-  phone: string,
-}
+  name: string;
+  netid: string;
+  email: string;
+  phone: string;
+};
 
 function deleteEntry(netid: string, driverList: Driver[]) {
   return driverList.filter((driver) => driver.netid !== netid);
@@ -30,14 +30,34 @@ function addDriver(newDriver: Driver, allDrivers: Driver[]) {
 
 const Table = () => {
   const [drivers, setDrivers] = useState([
-    { name: 'Driver1', netid: 'abc123', email: 'abc123@cornell.edu', phone: '111-111-1111' },
-    { name: 'Driver2', netid: 'asd23', email: 'asd23@cornell.edu', phone: '222-222-2222' },
-    { name: 'Driver3', netid: 'advs12', email: 'advs12@cornell.edu', phone: '333-333-3333' },
-    { name: 'Driver4', netid: 'bfd2', email: 'bfd2@cornell.edu', phone: '444-444-4444' },
+    {
+      name: 'Driver1',
+      netid: 'abc123',
+      email: 'abc123@cornell.edu',
+      phone: '111-111-1111',
+    },
+    {
+      name: 'Driver2',
+      netid: 'asd23',
+      email: 'asd23@cornell.edu',
+      phone: '222-222-2222',
+    },
+    {
+      name: 'Driver3',
+      netid: 'advs12',
+      email: 'advs12@cornell.edu',
+      phone: '333-333-3333',
+    },
+    {
+      name: 'Driver4',
+      netid: 'bfd2',
+      email: 'bfd2@cornell.edu',
+      phone: '444-444-4444',
+    },
   ]);
 
   function renderTableData(allDrivers: Driver[]) {
-    return allDrivers.map((driver, index) => {
+    return allDrivers.map((driver) => {
       const { name, netid, email, phone } = driver;
       return (
         <tr key={netid}>
@@ -46,10 +66,9 @@ const Table = () => {
           <td>{email}</td>
           <td>{phone}</td>
           <td>
-            <button
-              onClick={() => setDrivers(deleteEntry(netid, allDrivers))}>
+            <button onClick={() => setDrivers(deleteEntry(netid, allDrivers))}>
               Delete
-              </button>
+            </button>
           </td>
         </tr>
       );
@@ -66,9 +85,11 @@ const Table = () => {
             {renderTableData(drivers)}
           </tbody>
         </table>
-      </div >
+      </div>
       <div>
-        <Form onClick={(newDriver) => setDrivers(addDriver(newDriver, drivers))} />
+        <Form
+          onClick={(newDriver) => setDrivers(addDriver(newDriver, drivers))}
+        />
       </div>
     </>
   );
