@@ -1,5 +1,9 @@
 import { createContext } from 'react';
 
-const AuthContext = createContext({ logout: () => { } });
+const AuthContext = createContext({
+  logout: () => {
+    // do nothing
+  },
+});
 
 export default AuthContext;
