@@ -2,26 +2,26 @@ import dynamoose from 'dynamoose';
 import { Vehicle, VehicleType } from './vehicle';
 
 type Availability = {
-  startTime: string,
-  endTime: string,
+  startTime: string;
+  endTime: string;
 }
 
 type AvailabilityType = {
-  Mon?: Availability,
-  Tue?: Availability,
-  Wed?: Availability,
-  Thu?: Availability,
-  Fri?: Availability,
+  Mon?: Availability;
+  Tue?: Availability;
+  Wed?: Availability;
+  Thu?: Availability;
+  Fri?: Availability;
 }
 
 export type DriverType = {
-  id: string,
-  firstName: string,
-  lastName: string,
-  availability: AvailabilityType,
-  vehicle: VehicleType,
-  phoneNumber: string,
-  email: string,
+  id: string;
+  firstName: string;
+  lastName: string;
+  availability: AvailabilityType;
+  vehicle: VehicleType;
+  phoneNumber: string;
+  email: string;
 };
 
 const availability = {
